+++ config/app.go
@@ -116,17 +116,19 @@ const (
 
 // NotifierConfig holds configuration
 type NotifierConfig struct {
-	LogFile       string `mapstructure:"LOG_FILE"`
-	TagID         string `mapstructure:"TAG_ID"`
-	RepoName      string `mapstructure:"REPO_NAME"`
-	RepoOwner     string `mapstructure:"REPO_OWNER"`
-	Token         string `mapstructure:"TOKEN"`
-	TokenUser     string `mapstructure:"TOKEN_USER"`
-	PullRequestID int    `mapstructure:"PR_ID"`
-	DeleteComment bool   `mapstructure:"DELETE_COMMENT"`
-	Vcs           string `mapstructure:"VERSION_CONTROL_SYSTEM"`
-	Ci            string `mapstructure:"CI_SYSTEM"`
-	Url           string `mapstructure:"URL"`
+	LogFile         string `mapstructure:"LOG_FILE"`
+	TagID           string `mapstructure:"TAG_ID"`
+	RepoName        string `mapstructure:"REPO_NAME"`
+	RepoOwner       string `mapstructure:"REPO_OWNER"`
+	Token           string `mapstructure:"TOKEN"`
+	TokenUser       string `mapstructure:"TOKEN_USER"`
+	PullRequestID   int    `mapstructure:"PR_ID"`
+	DeleteComment   bool   `mapstructure:"DELETE_COMMENT"`
+	Vcs             string `mapstructure:"VERSION_CONTROL_SYSTEM"`
+	Ci              string `mapstructure:"CI_SYSTEM"`
+	Url             string `mapstructure:"URL"`
+	NoPostMode      bool   `mapstructure:"NO_POST_MODE"`
+	DisableCollapse bool   `mapstructure:"DISABLE_COLLAPSE"`
 }
 
 // Init will create default NotifierConfig with following priority
@@ -192,6 +194,9 @@ func createBindings() map[string]string {
 }
 
 func (c *NotifierConfig) validate() error {
+	if c.NoPostMode {
+		return nil
+	}
 	ci := viper.GetString("ci_system")
 	if c.PullRequestID == 0 && ci == CiCircleCi {
 		pr := PullRequest{}