+++ transform/transformer.go
@@ -23,6 +23,9 @@ type LogTransformer struct {
 	LogContent string
 	Logfile    string
 	TagID      string
+	NoPostMode bool
+	Vcs  string
+	DisableCollapse bool
 }
 
 // githubTemplate wrapper object to use go templating
@@ -32,6 +35,7 @@ type githubTemplate struct {
 	JobLink      string
 	Backticks    string
 	HeaderPrefix string
+	Collapsible  bool
 }
 
 // NewLogTransformer create new log transfer based on config.AppConfig
@@ -40,6 +44,9 @@ func NewLogTransformer(config *config.NotifierConfig) *LogTransformer {
 		LogContent: "",
 		Logfile:    config.LogFile,
 		TagID:      config.TagID,
+		NoPostMode: config.NoPostMode,
+		Vcs: config.Vcs,
+		DisableCollapse: config.DisableCollapse,
 	}
 }
 
@@ -106,16 +113,34 @@ func (t *LogTransformer) truncate() {
 func (t *LogTransformer) addHeader() {
 	templateContent := `
 {{ .HeaderPrefix }} {{ .TagID }} {{ .JobLink }}
+{{- if .Collapsible }}
+<details>
+<summary>Click to expand</summary>
+{{- end }}
+
 {{ .Backticks }}diff
 {{ .Content }}
 {{ .Backticks }}
+{{- if .Collapsible }}
+</details>
+{{- end }}
 `
+	collapsible := false
+	// only github and gitlab support collapsable sections
+	if t.Vcs == "github" || t.Vcs == "gitlab" {
+		collapsible = true
+	}
+	// can be disable by command line
+	if t.DisableCollapse {
+		collapsible = false
+	}
 	githubTemplate := &githubTemplate{
 		TagID:        t.TagID,
 		Content:      t.LogContent,
 		Backticks:    "```",
 		JobLink:      "",
 		HeaderPrefix: provider.HeaderPrefix,
+		Collapsible:  collapsible,
 	}
 	tmpl, err := template.New("githubTemplate").Parse(templateContent)
 	if err != nil {
@@ -133,11 +158,27 @@ func (t *LogTransformer) printFile() {
 	logrus.Infof("File contents: %s", t.LogContent)
 }
 
+// writeDiffFile is writing the transformed diff to file and appends .diff to filename.
+// Additionally, the diff is streamed to stdout
+func (t *LogTransformer) writeDiffFile() error {
+	if !t.NoPostMode {
+		return nil
+	}
+	filePath := t.Logfile + ".diff"
+	err := ioutil.WriteFile(filePath, []byte(t.LogContent), 440)
+	if err != nil {
+		return err
+	}
+	fmt.Print(t.LogContent)
+	return nil
+}
+
 // Process log file
 // 1. Clean any ANSI chars and XTERM color created from cdk diff command
 // 2. Transform additions and removals to markdown diff syntax
 // 3. Create unique message header
 // 4. truncate content if message is longer than GitHub API can handle
+// 5. write diff as file and to stdout when no-post-mode is activated
 func (t *LogTransformer) Process() {
 	err := t.readFile()
 	if err != nil {
@@ -147,4 +188,8 @@ func (t *LogTransformer) Process() {
 	t.transformDiff()
 	t.addHeader()
 	t.truncate()
+	err = t.writeDiffFile()
+	if err != nil {
+		logrus.Fatal(err)
+	}
 }